+++ src/app/member/[fridgeId]/page.tsx
@@ -11,9 +11,9 @@ import AddInventoryButton from "@/components/inventory/add-inventory-button";
 import { getInventories } from "@/lib/inventory";
 import { getShoppingList } from "@/lib/shopping-list";
 import AddPurchaseButton from "@/components/purchase/add-purchase-button";
-import { getPurchases } from "@/lib/purchase";
+import { getPurchases, getPurchasesUsers } from "@/lib/purchase";
 import PurchaseList from "@/components/purchase/purchase-list";
-import { getFridgeAccount } from "@/lib/fridge";
+import { PurchasesUserType } from "@/types/types";
 
 const FridgePage = async () => {
   const session = await getServerSession(nextAuthOptions);
@@ -29,8 +29,7 @@ const FridgePage = async () => {
   const inventories = await getInventories(fridgeId);
   const purchases = await getPurchases(fridgeId);
   const now = new Date();
-  const fridgeAccount = await getFridgeAccount(fridgeId);
-  const fridgeAccountUsers = fridgeAccount.userFridges
+  const purchasesUsers: PurchasesUserType[] = getPurchasesUsers(purchases);
   return (
     <div className="w-full md:flex md:flex-wrap md:gap-x-11 md:gap-y-12 max-md:space-y-6">
       <Box variant="rounded" className="md:order-1 md:flex-1">
@@ -69,7 +68,7 @@ const FridgePage = async () => {
           fridgeId={fridgeId}
           date={now}
           purchases={purchases}
-          users={fridgeAccountUsers}
+          users={purchasesUsers}
           headingStyle="max-md:text-center"
         />
       </Box>
