+++ src/app/member/[fridgeId]/purchases/page.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth";
 import PurchasesList from "@/components/purchase/purchases-list";
-import { getPurchases } from "@/lib/purchase";
+import { getPurchases, getPurchasesUsers } from "@/lib/purchase";
 import { redirect } from "next/navigation";
 import { nextAuthOptions } from "@/lib/next-auth/options";
 import Heading from "@/components/ui/heading";
@@ -29,22 +29,7 @@ const PurchasesPage = async () => {
       return 0;
     }
   });
-  const sortedPurchasesByUser: PurchaseType[] = [...purchases];
-  sortedPurchasesByUser.sort((first, second) => {
-    if (first.userId > second.userId) {
-      return -1;
-    } else if (second.userId > first.userId) {
-      return 1;
-    } else {
-      return 0;
-    }
-  });
-  const users: PurchasesUserType[] = [];
-  sortedPurchasesByUser.forEach((purchase, idx) => {
-    if (idx === 0 || purchase.userId != sortedPurchasesByUser[idx - 1].userId) {
-      users.push({ id: purchase.userId, name: purchase.user.name });
-    }
-  });
+  const users: PurchasesUserType[] = getPurchasesUsers(purchases);
   return (
     <>
       <Heading level={1} icon={FaFileLines}>
