+++ src/components/inventory/inventory-table.tsx
@@ -25,7 +25,7 @@ const InventoryTable = ({
   className,
   ...props
 }: InventoryTableProps) => {
-  const baseStyle = ""
+  const baseStyle = "";
   const [sortedInventories, setSortedInventories] =
     useState<InventoryType[]>(inventories);
   useEffect(() => {
@@ -94,60 +94,67 @@ const InventoryTable = ({
   };
   return (
     <>
-    {inventories.length === 0 ? <Paragraph className="px-4 py-2">在庫管理品は登録されていません</Paragraph>:
-    <Table className={cn(baseStyle,className)} {...props}>
-      <TableHead>
-        <TableRow>
-          <TableHeader className="text-left">
-            品名
-            <br />
-            <Button
-              variant="angle"
-              angle="up"
-              onClick={handleSortNameAscending}
-            />
-            <Button
-              variant="angle"
-              angle="down"
-              onClick={handleSortNameDescending}
-              className="ml-2"
-            />
-          </TableHeader>
-          <TableHeader className="sm:w-20 max-sm:w-15">
-            残数
-            <br />
-            <Button
-              variant="angle"
-              angle="up"
-              onClick={handleSortRemainingAscending}
-            />
-            <Button
-              variant="angle"
-              angle="down"
-              onClick={handleSortRemainingDescending}
-              className="ml-2"
-            />
-          </TableHeader>
-          <TableHeader className="w-10"></TableHeader>
-        </TableRow>
-      </TableHead>
-      <TableBody className="max-h-30">
-        {sortedInventories.map((inventory, idx) => (
-          <TableRow key={idx}>
-            <TableData>
-              {inventory.name}
-              <span className="text-xs text-gray pl-0.5">
-                ({categories[inventory.category]})
-              </span>
-            </TableData>
-            <TableData className="text-center">{inventory.remaining}</TableData>
-            <TableData className="text-center">
-              <EditInventoryButton inventory={inventory} />
-            </TableData>
-          </TableRow>
-        ))}
-      </TableBody>
-    </Table>}
+      {inventories.length === 0 ? (
+        <Paragraph className="px-4 py-2">
+          右上の「＋」ボタンを押して在庫管理品を追加してください。
+        </Paragraph>
+      ) : (
+        <Table className={cn(baseStyle, className)} {...props}>
+          <TableHead>
+            <TableRow>
+              <TableHeader className="text-left">
+                品名
+                <br />
+                <Button
+                  variant="angle"
+                  angle="up"
+                  onClick={handleSortNameAscending}
+                />
+                <Button
+                  variant="angle"
+                  angle="down"
+                  onClick={handleSortNameDescending}
+                  className="ml-2"
+                />
+              </TableHeader>
+              <TableHeader className="sm:w-20 max-sm:w-15">
+                残数
+                <br />
+                <Button
+                  variant="angle"
+                  angle="up"
+                  onClick={handleSortRemainingAscending}
+                />
+                <Button
+                  variant="angle"
+                  angle="down"
+                  onClick={handleSortRemainingDescending}
+                  className="ml-2"
+                />
+              </TableHeader>
+              <TableHeader className="w-10"></TableHeader>
+            </TableRow>
+          </TableHead>
+          <TableBody className="max-h-30">
+            {sortedInventories.map((inventory, idx) => (
+              <TableRow key={idx}>
+                <TableData>
+                  {inventory.name}
+                  <span className="text-xs text-gray pl-0.5">
+                    ({categories[inventory.category]})
+                  </span>
+                </TableData>
+                <TableData className="text-center">
+                  {inventory.remaining}
+                </TableData>
+                <TableData className="text-center">
+                  <EditInventoryButton inventory={inventory} />
+                </TableData>
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      )}
     </>
   );
 };
